requirements.txt
pandas==1.5.3
numpy==1.24.4
psycopg2-binary==2.9.6
faker==18.9.0
python-dotenv==1.0.0
